requirements.txt
streamlit
librosa
numpy
pandas
scipy
scikit-learn
xgboost
imblearn
joblib
pydub
openai-whisper
textblob
vaderSentiment
textstat
openpyxl
python-dotenv
torch
ffmpeg-python	
openai-whisper
SpeechRecognition